Refactor Generator component with AngularGenerator
