+++ src/main/webapp/app/springboot/primary/angular-generator/index.ts
@@ -1,0 +1,4 @@
+import AngularGeneratorComponent from './AngularGenerator.component';
+import AngularGeneratorVue from './AngularGenerator.vue';
+
+export { AngularGeneratorComponent, AngularGeneratorVue };
